requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
apsw==3.40.1.0
attrs==21.4.0
boto3==1.28.62
botocore==1.31.62
CacheControl==0.12.10
certifi==2023.07.22
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
contextlib2==21.6.0
croniter==1.2.0
distlib==0.3.3
distro==1.6.0
Flask==2.2.5
Flask-HTTPAuth==4.5.0
flask-restx==1.1.0
html5lib==1.1
idna==3.7
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
lockfile==0.12.2
MarkupSafe>=2.1.1
msgpack==1.0.2
ordered-set==4.0.2
packaging==20.9
pep517==0.12.0
progress==1.6
prometheus-client==0.13.1
pyhocon==0.3.59
pymongo==4.6.3
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.3
s3transfer==0.7.0
six==1.16.0
toml==0.10.2
tomli==1.2.2
urllib3==2.0.6
webencodings==0.5.1
Werkzeug==3.0.3
